feat(ai): reproduce Kratos issue 2761 about missing path variable
This commit adds a demo to reproduce Kratos Issue 2761, where
the `user_id` URL parameter is not encoded in the HTTP request.
It includes the necessary proto definitions, server and client
implementations, and a README with instructions to reproduce the issue.
Besides, it upgrades Kratos to v2.8.4 and other dependencies.

Signed-off-by: Ruins He <lyhypacm@gmail.com>